Update Animations.js to play animations at the set FPS regardless of runtime frame limit
Changing the editor FPS after setting the FPS of an animation will cause the animation to change speed accordingly. (ex. changing from 30 to 60 will make the anim play twice as fast).

Some games have the FPS set to 0 (VSync), so I thought this would help the animaitons be more accurate.